cast pages_count to integer from float

+++ app.py
@@ -22,7 +22,7 @@ class reports:
     def GET(self):
         page_size = 10
         reports_count = db.get_reports_count()
-        pages_count = math.ceil(reports_count / (page_size * 1.0))
+        pages_count = int(math.ceil(reports_count / (page_size * 1.0)))
         current_page = int(web.input(page=1).page)
         if current_page > pages_count:
             web.seeother('/reports')
